Allow for distinct build command directory and code directory
When auto-annotating a single target in a large repo, for some
build systems (namely buck), the build command must be run from
the repo root, whereas the code to be annotated will live in a
sub-directory.

We allow specifying an optional config parameter (`REPO_ROOT_PATH`),
which will be used as the directory from which to run the build
command (if present), separate from the parameter specifying the
source code location.

If this parameter is missing, run.py should maintain its existing
behavior.